+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.player.Hand;
 import fixture.CardFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
