+++ src/test/java/blackjack/domain/player/HandTest.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardShape;
 import fixture.HandFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
