+++ src/test/java/blackjack/domain/player/PlayersTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import fixture.PlayerFixture;
 import org.junit.jupiter.api.DisplayName;
